third_party/blink/web_tests/fast/forms/select/select-show-picker.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<body>
<select id="select">
  <option>1</option>
</select>
</body>
<script>
promise_test(async (t) => {
  await test_driver.bless("show select picker");
  select.showPicker();
  assert_true(internals.isSelectPopupVisible(select), "Picker is opened.");
}, "Select: showPicker() should show popup");
</script>
